Add Insert And Get from Local DataBase

+++ app/src/main/java/Interfaces/OnCheckBoxClickListener.kt
@@ -3,5 +3,5 @@ package Interfaces
 import android.icu.text.Transliterator.Position
 
 interface OnCheckBoxClickListener {
-    fun onClickCheckBox(isChecked: Boolean, position: Int)
+    fun onClickCheckBox(isChecked: Boolean, position: Long)
 }
